extract parse_url
